:recycle: refactor - Profile.tsx 페이지, Footer 컴포넌트 반응형 디자인 작업
작업 내용
- HeroSection.tsx 위치 이동 (Layout 폴더로 이동)
- FooterColumn 기본 사이즈 ~ xs 사이즈 최대 너비 조정
- ContactUs.tsx md, lg 레이아웃 수정

기본 디자인 작업
- PC - ~ xs(~ 365px)
- PC - lg(1024px)
- PC - xl(1280px)

작업 디바이스
- Galaxy Z Fold 5 (344 * 882)
- 반응형 (1055 * 1161)
- 반응형 (1339 * 1161) (CL: 471)
