+++ src/features/profile/pages/ContactUs.tsx
@@ -41,13 +41,13 @@
     return (
       <SectionLayout
         title="CONTACT US"
-        className="mx-auto px-4 sm:px-6 md:px-8 lg:px-10 xl:px-12 flex max-w-6xl flex-col items-center justify-center py-12 sm:py-16 lg:py-20"
+        className="mx-auto px-4 sm:px-6 md:px-8 lg:px-5 flex max-w-6xl xl:max-w-7xl flex-col items-center justify-center py-12 sm:py-16 lg:py-20"
         titleClassName="py-6 text-3xl lg:text-4xl font-semibold text-center"
         fullHeight={false}
       >
         {/* 연락처 정보 */}
-        <div className="grid w-full grid-cols-1 md:mt-10 md:grid-cols-2 lg:grid-cols-5 gap-8 md:gap-12">
-          <ul className="mx-auto flex max-w-sm flex-col gap-6 lg:col-span-2">
+        <div className="grid w-full grid-cols-1 lg:mt-10 md:grid-cols-1 lg:grid-cols-5 gap-8 lg:gap-12">
+          <ul className="mx-auto max-w-sm md:max-w-4xl xl:max-w-5xl flex flex-col md:grid md:grid-cols-2 gap-6 lg:col-span-2 lg:flex lg:flex-col">
             {contactInfo.map((item, index) => (
               <ListCard
                 key={index}
@@ -55,20 +55,15 @@
                 type="_base"
                 icon={item.icon}
                 description={item.description}
-                size="w-full h-full"
-              />
+                size="w-full h-full" />
             ))}
           </ul>
 
           {/* 이미지 영역 */}
-          <div
-            id="map"
-            className="mx-auto max-w-sm w-full h-64 md:h-full lg:col-span-3 lg:mx-0 lg:max-w-full lg:h-full  rounded-2xl pt-10"
-          >
+          <div id="map" className="w-full h-64 mx-auto max-w-sm md:max-w-4xl md:h-[600px] lg:col-span-3 lg:mx-0 lg:max-w-full lg:h-full  rounded-2xl pt-10">
             <NaverMap
               latitude={37.48771788371748}
-              longitude={127.00855064427329}
-            />
+              longitude={127.00855064427329} />
           </div>
         </div>
       </SectionLayout>
