+++ src/layout/HeroSection.tsx
